+++ src/Factory.ts
@@ -1,18 +1,76 @@
 import { Node } from './Node';
+import { GetSet } from './types';
 import { Util } from './Util';
 import { getComponentValidator } from './Validators';
 
-const GET = 'get',
-  SET = 'set';
+const GET = 'get';
+const SET = 'set';
+
+/**
+ * Enforces that a type is a string.
+ */
+type EnforceString<T> = T extends string ? T : never;
+
+/**
+ * Represents a class.
+ */
+type Constructor = abstract new (...args: any) => any;
+
+/**
+ * An attribute of an instance of the provided class. Attributes names be strings.
+ */
+type Attr<T extends Constructor> = EnforceString<keyof InstanceType<T>>;
+
+/**
+ * A function that is called after a setter is called.
+ */
+type AfterFunc<T extends Constructor> = (this: InstanceType<T>) => void;
+
+/**
+ * Extracts the type of a GetSet.
+ */
+type ExtractGetSet<T> = T extends GetSet<infer U, any> ? U : never;
+
+/**
+ * Extracts the type of a GetSet class attribute.
+ */
+type Value<T extends Constructor, U extends Attr<T>> = ExtractGetSet<
+  InstanceType<T>[U]
+>;
+
+/**
+ * A function that validates a value.
+ */
+type ValidatorFunc<T> = (val: ExtractGetSet<T>, attr: string) => T;
+
+/**
+ * Extracts the "components" (keys) of a GetSet value. The value must be an object.
+ */
+type ExtractComponents<T extends Constructor, U extends Attr<T>> = Value<
+  T,
+  U
+> extends Record<string, any>
+  ? EnforceString<keyof Value<T, U>>[]
+  : never;
 
 export const Factory = {
-  addGetterSetter(constructor, attr, def?, validator?, after?) {
+  addGetterSetter<T extends Constructor, U extends Attr<T>>(
+    constructor: T,
+    attr: U,
+    def?: Value<T, U>,
+    validator?: ValidatorFunc<Value<T, U>>,
+    after?: AfterFunc<T>
+  ): void {
     Factory.addGetter(constructor, attr, def);
     Factory.addSetter(constructor, attr, validator, after);
     Factory.addOverloadedGetterSetter(constructor, attr);
   },
-  addGetter(constructor, attr, def?) {
-    const method = GET + Util._capitalize(attr);
+  addGetter<T extends Constructor, U extends Attr<T>>(
+    constructor: T,
+    attr: U,
+    def?: Value<T, U>
+  ) {
+    var method = GET + Util._capitalize(attr);
 
     constructor.prototype[method] =
       constructor.prototype[method] ||
@@ -22,15 +80,26 @@ export const Factory = {
       };
   },
 
-  addSetter(constructor, attr, validator?, after?) {
-    const method = SET + Util._capitalize(attr);
+  addSetter<T extends Constructor, U extends Attr<T>>(
+    constructor: T,
+    attr: U,
+    validator?: ValidatorFunc<Value<T, U>>,
+    after?: AfterFunc<T>
+  ) {
+    var method = SET + Util._capitalize(attr);
 
     if (!constructor.prototype[method]) {
       Factory.overWriteSetter(constructor, attr, validator, after);
     }
   },
-  overWriteSetter(constructor, attr, validator?, after?) {
-    const method = SET + Util._capitalize(attr);
+
+  overWriteSetter<T extends Constructor, U extends Attr<T>>(
+    constructor: T,
+    attr: U,
+    validator?: ValidatorFunc<Value<T, U>>,
+    after?: AfterFunc<T>
+  ) {
+    var method = SET + Util._capitalize(attr);
     constructor.prototype[method] = function (val) {
       if (validator && val !== undefined && val !== null) {
         val = validator.call(this, val, attr);
@@ -45,12 +114,13 @@ export const Factory = {
       return this;
     };
   },
-  addComponentsGetterSetter(
-    constructor,
-    attr: string,
-    components: Array<string>,
-    validator?: Function,
-    after?: Function
+
+  addComponentsGetterSetter<T extends Constructor, U extends Attr<T>>(
+    constructor: T,
+    attr: U,
+    components: ExtractComponents<T, U>,
+    validator?: ValidatorFunc<Value<T, U>>,
+    after?: AfterFunc<T>
   ) {
     const len = components.length,
       capitalize = Util._capitalize,
@@ -59,7 +129,7 @@ export const Factory = {
 
     // getter
     constructor.prototype[getter] = function () {
-      const ret = {};
+      const ret: Record<string, any> = {};
 
       for (let n = 0; n < len; n++) {
         const component = components[n];
@@ -76,7 +146,7 @@ export const Factory = {
       const oldVal = this.attrs[attr];
 
       if (validator) {
-        val = validator.call(this, val);
+        val = validator.call(this, val, attr);
       }
 
       if (basicValidator) {
@@ -106,8 +176,11 @@ export const Factory = {
 
     Factory.addOverloadedGetterSetter(constructor, attr);
   },
-  addOverloadedGetterSetter(constructor, attr) {
-    const capitalizedAttr = Util._capitalize(attr),
+  addOverloadedGetterSetter<T extends Constructor, U extends Attr<T>>(
+    constructor: T,
+    attr: U
+  ) {
+    var capitalizedAttr = Util._capitalize(attr),
       setter = SET + capitalizedAttr,
       getter = GET + capitalizedAttr;
 
@@ -121,7 +194,12 @@ export const Factory = {
       return this[getter]();
     };
   },
-  addDeprecatedGetterSetter(constructor, attr, def, validator) {
+  addDeprecatedGetterSetter<T extends Constructor, U extends Attr<T>>(
+    constructor: T,
+    attr: U,
+    def: Value<T, U>,
+    validator: ValidatorFunc<Value<T, U>>
+  ) {
     Util.error('Adding deprecated ' + attr);
 
     const method = GET + Util._capitalize(attr);
@@ -139,7 +217,10 @@ export const Factory = {
     });
     Factory.addOverloadedGetterSetter(constructor, attr);
   },
-  backCompat(constructor, methods) {
+  backCompat<T extends Constructor>(
+    constructor: T,
+    methods: Record<string, string>
+  ) {
     Util.each(methods, function (oldMethodName, newMethodName) {
       const method = constructor.prototype[newMethodName];
       const oldGetter = GET + Util._capitalize(oldMethodName);
@@ -161,7 +242,7 @@ export const Factory = {
       constructor.prototype[oldSetter] = deprecated;
     });
   },
-  afterSetFilter(this: Node) {
+  afterSetFilter(this: Node): void {
     this._filterUpToDate = false;
   },
 };