+++ src/filters/Sepia.ts
@@ -12,17 +12,13 @@ import { Filter } from '../Node';
  * node.filters([Konva.Filters.Sepia]);
  */
 export const Sepia: Filter = function (imageData) {
-  let data = imageData.data,
-    nPixels = data.length,
-    i,
-    r,
-    g,
-    b;
+  const data = imageData.data,
+    nPixels = data.length;
 
-  for (i = 0; i < nPixels; i += 4) {
-    r = data[i + 0];
-    g = data[i + 1];
-    b = data[i + 2];
+  for (let i = 0; i < nPixels; i += 4) {
+    const r = data[i + 0];
+    const g = data[i + 1];
+    const b = data[i + 2];
 
     data[i + 0] = Math.min(255, r * 0.393 + g * 0.769 + b * 0.189);
     data[i + 1] = Math.min(255, r * 0.349 + g * 0.686 + b * 0.168);